requirements.txt
alembic==1.9.0
azure-functions
autopep8==2.0.1
azure-common==1.1.28
azure-communication-email==1.0.0b1
azure-core==1.26.1
azure-mgmt-core==1.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
msrest==0.7.1
numpy==1.23.5
oauthlib==3.2.2
pandas==1.5.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
pyodbc==4.0.35
